+++ frontend/src/components/pages/TeacherPage.tsx
@@ -4,7 +4,7 @@ import { FormProvider, useForm } from "react-hook-form";
 import { ClassroomForm } from "../../types/ClassroomTypes";
 import Logout from "../auth/Logout";
 import RefreshCredentials from "../auth/RefreshCredentials";
-import AddClassroomModal from "../user-management/AddClassroomModal";
+import AddClassroomModal from "../user-management/student/AddClassroomModal";
 
 const defaultValues = {
   className: "",
