+++ frontend/src/components/pages/admin/DisplayAssessmentsPage.tsx
@@ -21,16 +21,6 @@ import FilterMenu from "../../common/table/FilterMenu";
 import SearchBar from "../../common/table/SearchBar";
 import AssessmentsTable from "../../assessments/AssessmentsTable";
 
-// const ErrorState = (): React.ReactElement => (
-//   <VStack spacing={6} textAlign="center">
-//     <AlertIcon />
-//     <Text textStyle="paragraph" color="blue.300">
-//       The data has not loaded properly. Please reload the page or contact Jump
-//       Math.
-//     </Text>
-//   </VStack>
-// );
-
 const getAssessments = (assessment: AssessmentTypes) => {
   return {
     status: assessment.status,
